sentiment_analysis/pyproject.toml
[project]
name = "absa"
version = "0.1.0"
description = "Hierarchical ABSA with BERTopic aspect discovery on Reddit product discourse"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # CLI
    "typer>=0.12",
    "rich>=13.0",

    # Reddit data collection
    "praw>=7.7",

    # NLP preprocessing
    "spacy>=3.7",

    # Topic modeling
    "bertopic>=0.16",
    "sentence-transformers>=3.0",
    "umap-learn>=0.5",
    "hdbscan>=0.8",

    # Transformer ABSA
    "transformers>=4.40",
    "torch>=2.2",
    "pyabsa>=2.3",

    # Lexicon baseline
    "vaderSentiment>=3.3",

    # LLM (Gemini)
    "google-generativeai>=0.5",

    # Graph + aggregation
    "networkx>=3.3",
    "numpy>=1.26",
    "scipy>=1.13",

    # Config and persistence
    "pyyaml>=6.0",
    "python-dotenv>=1.0",

    # Visualization (optional HTML reports)
    "plotly>=5.20",
]

[project.scripts]
absa = "absa.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/absa"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
